% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_BaySIR.R
\docType{package}
\name{BaySIR}
\alias{BaySIR}
\title{BaySIR}
\description{
The BaySIR package includes two major functions, which can be used for 
estimating the time-varying effective reproduction number of COVID-19 in a
specific country/region, and predicting future case counts. 
Data of daily confirmed cases are needed. 
The two functions are:
\describe{
\item{\code{\link{BaySIR_MCMC}}}{}
\item{\code{\link{BaySIR_predict}}}{}
}
Use \code{?Function_Name} or \code{help(Function_Name)} to retrieve the 
documentation for a specific function. E.g., \code{?BaySIR_MCMC}.
}
\section{Author}{

Tianjian Zhou, \email{tjzhou95@gmail.com}
}

\seealso{
\code{\link{BaySIR_MCMC}} for MCMC sampling, 
  \code{\link{BaySIR_predict}} for posterior prediction.
}
