% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_BaySIR.R
\name{BaySIR_MCMC}
\alias{BaySIR_MCMC}
\title{BaySIR MCMC sampling}
\usage{
BaySIR_MCMC(
  B,
  I_D_0,
  N,
  confirmed_cases_cum = NULL,
  X = NULL,
  Y = NULL,
  kappa = 1,
  niter = 1000,
  burnin = 10000,
  thin = 20,
  Delta = (1.5)^(0:9)
)
}
\arguments{
\item{B}{A length \code{T + 1} vector of daily new confirmed cases 
\code{B[0], ..., B[T]}. \code{B[t]} represents the increment in
confirmed cases between day \code{t} and day \code{t+1}.}

\item{I_D_0}{The total number of confirmed cases on day 0.}

\item{N}{The population size.}

\item{confirmed_cases_cum}{An optional length \code{T + 2} vector of cumulative 
confirmed case counts. If \code{B} and \code{I_D_0} have already been 
specified, then \code{confirmed_cases_cum} will be ignored. If not both 
\code{B} and \code{I_D_0} are specified and \code{confirmed_cases_cum}
is supplied, then \code{confirmed_cases_cum}
will be used to calculate \code{B} and \code{I_D_0}.}

\item{X}{A \code{(T + 1) * Q} matrix, covariates related to the disease 
transmission rate. Default is an intercept term plus a time trend, 
\code{X[t, ] = (1, t)}.}
}
\value{
A list of the following:
\describe{
\item{\code{MCMC_spls}}{Again, a list of MCMC samples for the parameters.
\itemize{
  \item \code{R_eff_spls} A \code{(T + 1) * niter} matrix, MCMC samples of the 
                      effective reproduction number. The l-th column,
                      \code{R_eff_spls[ , l]}, represent the posterior samples of 
                      the effective reproduction numbers for \code{T + 1} days at
                      the l-th MCMC iteration.
}}
\item{\code{MCMC_summary}}{Posterior summaries for the parameters.}
}
}
\description{
Implementing the parallel-tempering Markov chain Monte Carlo 
(PTMCMC) sampling from the posterior distribution of the parameters.
}
\examples{

}
